rename to WindowTemplate
