+++ src/components/templates/WindowTemplate.vue
